tests/wpt/tests/html/canvas/element/path-objects/2d.path.stroke.overlap.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.path.stroke.overlap</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.path.stroke.overlap</h1>
<p class="desc">Stroked subpaths are combined before being drawn</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="2d.path.stroke.overlap.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("Stroked subpaths are combined before being drawn");
_addTest(function(canvas, ctx) {

  ctx.fillStyle = '#000';
  ctx.fillRect(0, 0, 100, 50);

  ctx.strokeStyle = 'rgba(0, 255, 0, 0.5)';
  ctx.lineWidth = 50;
  ctx.moveTo(0, 20);
  ctx.lineTo(100, 20);
  ctx.moveTo(0, 30);
  ctx.lineTo(100, 30);
  ctx.stroke();

  _assertPixelApprox(canvas, 50,25, 0,127,0,255, 1);

});
</script>
